(#33) Input side converter rejection solution when output side power is invalid due to bias change

+++ aspects/electrical/Batt/GunnsElectBattery.cpp
@@ -175,7 +175,7 @@ void GunnsElectBattery::initialize(GunnsElectBatteryConfigData&  configData,
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 void GunnsElectBattery::allocateArrays()
 {
-    TS_NEW_CLASS_ARRAY_EXT(mCells, mNumCells, GunnsElectBatteryCell, (), std::string(mName) + ".mCell");
+    TS_NEW_CLASS_ARRAY_EXT(mCells, static_cast<int>(mNumCells), GunnsElectBatteryCell, (), std::string(mName) + ".mCell");
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
